Update tests
- Update README
- Add `.env.e2e`
- Add `nvmrc`
- Some renamings: `tests` -> `describe/it`, `integration` -> `e2e` etc.
- Remove `jest-serial-runner`
