+++ .babelrc
@@ -1,7 +1,7 @@
 {
-  "presets": [
-    "@babel/preset-env"
-  ],
+  "presets": [[
+    "@babel/preset-env", {"targets": {"node": 14}}
+  ]],
   "plugins": [
     [
       "@babel/plugin-transform-runtime",
