+++ .e2e.env.sample
@@ -1,0 +1,2 @@
+THOR_PUB_KEY=XYZ
+THOR_ADDRESS=thorxyz
