+++ jest.config.js
@@ -15,17 +15,12 @@ module.exports = {
   // The directory where Jest should output its coverage files
   coverageDirectory: "coverage",
 
-  // A list of paths to directories that Jest should use to search for files in
-  roots: ["<rootDir>"],
-
-  runner: "jest-serial-runner",
-
   // The test environment that will be used for testing
   testEnvironment: "node",
 
   // The glob patterns Jest uses to detect test files
   testMatch: [
-    "**/__tests__/**/*.[jt]s?(x)",
+    "**/tests/**/*.[jt]s?(x)",
     "**/?(*.)+(spec|test).[tj]s?(x)",
     "**/?(*.)+(ispec|test).[tj]s?(x)",
   ],
