+++ jest.config.unit.js
@@ -1,7 +1,7 @@
 const config = require("./jest.config");
 
 // Override default configuration
-config.testMatch = ["**/__tests__/**/*.[jt]s?(x)", "**/?(*.)+(spec|test).[tj]s?(x)"];
+config.testMatch = ["**/tests/unit/**/*.[jt]s?(x)"];
 
 // eslint-disable-next-line no-console
 console.log("RUNNING UNIT TESTS");
