+++ package.json
@@ -42,9 +42,9 @@
     "bip39": "3.0.4",
     "core-js": "^3.15.2",
     "crypto-js": "4.0.0",
+    "dotenv": "^10.0.0",
     "index.js": "^0.0.3",
-    "jest": "^27.0.4",
-    "jest-serial-runner": "^1.1.0",
+    "jest": "^27.0.6",
     "prettier": "^2.3.2",
     "secp256k1": "^4.0.2",
     "vue": "^2.6.14",
@@ -53,7 +53,7 @@
   "scripts": {
     "build": "babel src --out-dir dist",
     "test:unit": "jest -c jest.config.unit.js",
-    "test:integration": "jest -c jest.config.integration.js --runInBand --detectOpenHandles",
+    "test:e2e": "jest -c jest.config.e2e.js --runInBand --detectOpenHandles",
     "serve": "openssl req -nodes -new -x509 -config certs/openssl.cnf -keyout certs/server.key -out certs/server.cert; openssl x509 -in certs/server.cert -out certs/cert.pem -outform PEM; vue-cli-service serve vue_example/main.js"
   },
   "moduleDirectories": [
